refactor: 공통 컴포넌트 common으로 이동
